requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-groq
langchain-pinecone
langchain-huggingface
sentence-transformers
pinecone-client
pydantic